Attachment API rework (#114)
A massive rework for our attachment and enablement paradigms for web controls targetted for KeymanWeb 10.0, merged from #114.  This also more closely ties the control paths used for page initialization with those used by MutationObservers for dynamically-added elements in an aim to improve internal consistency and facilitate new and future extensions to KMW functionality.