Got all the typing working in config windows and most link navigation
Still need to investigate why some links to help.keyman.com/... don't work in download window. Also, this check-in contains some commmented-out code to try to get mouse pointer to change over hyperlinks. Need to get it working or revert (possibly need to report as a separate issue). Also, it would be nice for there to be some user feedback when loading pages, which can be quite slow at times.