[Mac] [WIP] #769 - Added code to using Fabric with Crashlytics
This commit does not include the project and plist changes necessary to build/work because they have the APIKey in them and we don't want to publish that. Also, I haven't completely worked out where the frameworks should go or how that plumbing should be hooked up to work well on the build machines and for other developers.