Ensure keyboards list has an entry
There are some crash reports coming from `KeyboardPickerActivity.java` because the keyboards list is null (file access error). When the app tries to display the installed keyboards list, this results in the crash.

`getKeyboardsList()` already checks the existence of the `keyboards_list.dat` file, so some of that code is also redundant.