Implement looking up a wordform in a trie.
This trie implementation allows for word forms to be converted to keys.
Since keys do not have to have the exact same spelling, capitalization,
or even Unicode normalization form as the desried word form, this
enables a somewhat search for word forms based on prefixes.

The implementation was at one point Conrad Irwin's trie-ing,
<https://github.com/ConradIrwin/trie-ing/tree/9f1d2e08580bfd0bbf9d0f32db8ee450e1e4e9ad?
but has been heavily modfied since.