feat(developer): debugger menu to switch models
Fixes #2402. This adds a menu to switch models in the web debugger.
It is likely that most users will have only a single model loaded but
we need this to complete the circle. It also gives a little more
visibility in the model debugging process to validate that you are
testing the expected model.