feat: debug settings manager
Introduces a new app that lists all available
debug settings in Keyman Desktop and Developer
with a single unified UI to manage them. This
comes because I (as lead dev) am always forgetting
what the various keys and values are and need a
better way to manage them.

For now, this is alpha-quality and I do not intend
to deploy it with 13.0