fix(windows): '(Start Product)' shortcuts
Fixes #2467. This is a two part fix. First, the actual shortcuts
now work. Secondly, because the shortcut points to kmshell.exe,
Windows 10 Start Menu will hide it because it is identical to
the Keyman Desktop shortcut. So we add a uniquifier to un-hide
it. A bit silly, but makes sense in this instance.