fix(android): Revert KMManager.onDestroy
Fixes #2539

During #2326, `KMManager.onDestroy()` was updated to set `InAppKeyboard` and `SystemKeyboard` to null. While good in practice, this created  an issue when closing Keyman as a System Keyboard. Lingering Javascript calls in the WebView would crash with the null keyboards.

This reverts `onDestroy()` to how it has been since before 2.8 stable