fix(ci): is-build-required failed on missing branch
When the remote branch was deleted before the script ran, e.g. for
an automatically merged PR, this script would fail with error 128 from
git:

fatal: ambiguous argument 'origin/master...origin/auto/version-master-14.0.9':
unknown revision or path not in the working tree.

This handles error 128 as a signal to cancel the build rather than to
fail the build. It is possible that git would return error 128 for
other situations, but I cannot currently see any significant situations
where that would arise.