fix(ci): refactor to build based on central config
We have setup a single build configuration Keyman:Test that is triggered
when VCS changes are detected. This looks at the changes in the PR and
determines which projects need to be built. This is cleaner than the
current mess of configurations we have in TeamCity, because the
definitions are in just one place and far easier to maintain
consistently. This will also simplify all the pull request test
configurations because they won't need to be testing for safety or
other boilerplate as they will only ever be triggered by Keyma:Test
which has already run the safety tests.

The run-required-test-builds.sh script contains the logic; the data is
basically all in trigger-definitions.sh.