fix(common/core): buffer overrun in context api
A static buffer for UTF-8 encoded characters used in debugging and test
logs did not provide enough space for SMP characters (require 6 bytes
rather than 4). This was used only for test cases and not in production
code -- outputting JSON text.

This should fix the failing keyboardprocessor test builds on Windows and
macOS.