fix(developer/compiler): package: version mismatch
Fixes #2555.

The package compiler would give a misleading error when it was verifying
keyboard version numbers if one of the files could not be found; it
would say the keyboard file versions (between .kmx and .js) mismatched
instead of complaining about the missing file.

This reduces this error to a warning, allowing the verification to
continue, and the resulting log shows that the file is missing and
fails the build more logically.