fix(developer/ide): model debugger mime type mismatch
Fixes #2630.

The MIME type for .js was being read from the registry, which on some
systems has been corrupted to text/plain, which causes Chrome to abort
loading the script. The fix is to avoid loading the system registered
MIME types, which we don't need anyway, because all the MIME types we
are using are in the default hard-coded list.