refactor(mac): initial steps of input pathway
The handleEvent function has somewhat convoluted logic around emitting
the final transform to the client application. This starts to refactor
that code to move all client interaction to a single point at the end
of the event sequence.

Next phase will be to update delete and insert events to turn them into
a single action where possible.