chore(windows): improve build script tests
Make sure that our prerequisites are in place *before* starting a
release build. Note: doesn't currently check that certificates are
valid so that's a future test we could include.

One day it'd be nice to remove dependency on 7-zip: perhaps we won't
need it once we have better crash reporting.