feat(windows): sentrytool app
This app prepares Delphi executables for PDB-style
symbol information and creates a breakpad .sym file
from a corresponding .map file. x86 only at this point
but will work on x64 later. The differences are likely
to be minor. While this does not generate a .pdb file,
sentrytool could be extended to do this. So far, this
is just sentrytool; the build changes to use sentrytool
will come next.