fix(android): Clarify label that shows "Get Started" on startup
PR #2751 cleaned up UI strings, changing the label from `Show "Get Started" on startup` to `Show this menu on startup`. This change was fine for "Get Started" menu, but the label is confusing when reused on the Keyman Settings menu.

This PR reverts the label (while formatting the string to include the localized string for "Get Started".