fix(common): Fix meson.build script
`VERSION.md` is in the root of the source directory. However, when
building a Linux package from a source package we can't access files
from the root directory. Prior to creating the source package
`VERSION.md` gets copied to `common/core/desktop` so that it is
available in the source package. However, this hack broke the
meson build everywhere else unless `VERSION.md` got copied manually.

This change uses `VERSION.md` from the root directory if it is
available and otherwise assumes that it exists in the same directory
as the `meson.build` script. This change also switches to using
scripts to get the version number rather than doing it in the meson
build itself because that gives more flexibility. It is important to
note that the Windows batch file has to be listed first in
`find_program` because Windows might detect `getversion.sh` and
tries to run it with bash but this fails because of the different
directory separators.