feat(linux): Make two windows modal dialogs
This change modifies DownloadKmpWindow and InstallKmpWindow so that
they are modal dialogs. This will cause them to show up in front of
the km-config window and will prevent accidentally open multiple
download windows.

This change also fixes an unused method in ibus_util that was using
a variable that got renamed previously.