feat(windows): Keyman Setup online bootstrap
This starts the refactoring of Keyman Setup for Windows
for a more online-first bootstrap approach for bundling.
For users who are offline, the traditional bundled approach
works, as does throwing package files into the same folder
as setup.exe. This makes the whole installer much simpler
to deploy and test.

At this stage, there are a lot of rough edges, so need to
refine these, including setup strings, language choices,
and error and logging management.