fix(developer): Touch font size should be string
Fixes #3487.

The fontsize property for touch layout elements was treated as
an integer by Developer but should be a string (as it might contain
units such as 'em').

To help with back-compat, updated the GetValue string overload
to ignore value type and return the stringified form, regardless of
the actual type, which should always work.