fix(windows): show version with tag in setup
Fixes #3680.

Note that because the installer version may differ from the Keyman
version that will actually be installed, the installer version number is
a bit hidden away in the Options dialog, to try and reduce potential for
confusion.

For example, if a user already has the installer downloaded, but a new
version of Keyman is available, then by default the installer will try
and download the new version rather than install the current version.