feat(windows): kmconfig console app
This app uses the new `TKeymanSettings` class to present a standardised
interface for all Keyman settings -- debug, development and
compatibility. This is a console app, intended for use in a variety of
scenarios, including development tools, in the future.

Settings can be set, exported/imported to/from json, reset, shown.

Important in this PR is establishing the command line parameter
consistency and usage into the future.