chore(common): fixup missing history
Discovered that some history entries were missing. This was because we
were calling `--first-parent` on `git log` in fixupHistory.ts. (I will
fix that in the next commit.)

I manually constructed the remaining history by hacking up
fixupHistory.ts to pull together the additional entries. A bit scary but
it worked.

For ref, I threw together the following PHP to get the list of missing
PRs:

```
<?php
  $foo = shell_exec("git log --merges --oneline --decorate=short master...release-14.0.3-alpha | grep -E \"#|-alpha\"");
  $foo = explode("\n", $foo);

  $version = '14.0.187';

  $history = file_get_contents('HISTORY.md');

  foreach($foo as $line) {
    $PR = null;
    if(preg_match("/#(\d\d\d\d)/", $line, $matches)) {
      // Is this entry a PR
      if(!preg_match("/keymanapp\/auto\/version-master/", $line)) {
        // Skip automatic version upgrades
        $PR = $matches[1];
      }
    }
    if(preg_match('/\(tag: release-(\d+\.\d+\.\d+)-alpha\)/', $line, $matches)) {
      // Was there a version tag here?
      $version = $matches[1];
    }
    if($PR) {
      if(strpos($history, "(#$PR)") === false ) {
        echo " $version: $PR [$line]\n";
      }
    }
  }
?>
```

And I munged up fixupHistory.ts to just lookup each PR returned from
that report and splice it into HISTORY.md. That hackery is so awful I am
not sharing it here.