fix(web): sporadic blank keyboard on Android
Fixes #3996.

This is a tentative fix: it merely prevents `_Hide` from having any effect when in embedded mode. Tested on Android and could not force any issues.

However, it addresses the symptom and is probably sufficient but does not actually fix the root issue of the problem, which is that the `_Hide` method should never be called for embedded mode, as keyboard visibility is controlled by the host Keyman Engine on the device. Digging out those code paths is a much bigger job, and as the keyboard continues to function correctly if the effect of the hide (opacity=0) is undone by hand using web inspector, I am reasonably confident that this won't break other things.