fix(windows): remove obsolete option 'Switch to OSK/Help'
Fixes #4113.

The option "Switch to On Screen Keyboard/Help automatically when a
keyboard is selected" is obsolete and we should remove it from the UI.

Note that this does not remove the API surface, just hides the option
from view, as there may be apps relying on the API value.

Not touching help just now as that will be addressed in a help update
in the future.