chore(windows): sentrytool param file support
The maximum command line length in Windows is 2047 characters. In Delphi
it appears to be 1024 characters. In debugging, I hit that 1024
character limit which was pretty annoying.

So this patch allows us to use an `@file` parameter to put command line
parameters into a text file and read from that instead. It's probably a
little hacked together and could be turned into a class that we could
usefully reuse in all our other command line apps.

I looked around for a generic Delphi solution but did not find an
existing one that was any good.

Low priority, not required for beta, but putting it up because I've done
the work already.