fix(android): Popup misalignments and compatibility with WeChat, Telegram
Fixes #4019.
Rolls back #3009 and fixes base issue in alternative manner.

The core issue here was twofold:

1. We needed to set the window size to the whole screen in order for the
   popups to visible and touchable. We implement this in
   `onConfigureWindow`.
2. Once our window size is the whole screen, we need to tell the apps
   which part of the window is our keyboard UI, in `onComputeInsets`.
   This is the only part of the window which impacts the layout of the
   underlying app. We had not set the `visibleTopInsets` property and
   this caused the WeChat problems.

Finally, in order for this to work, we roll back the original fix
in #3009 as that one breaks the full-screen layout requirement.

Tested with a wide range of apps on Android 10, including WeChat and
Telegram. If @kienwt, @ericjding have capacity, would love to know if
this fix works for them also.

Note that this fix requires changes to consumers of Keyman Engine, so
sample and FV Keyboards updated accordingly.