fix(web): ensure only one keyboard stylesheet
As a new instance of `VisualKeyboard` may be setup, we can lose the
existing reference to the custom stylesheet for the keyboard, leaving
multiple stylesheets in the document. Making the stylesheet a static
member of `VisualKeyboard` is not the only way to resolve this, but it's
fairly straightforward.

Happy to entertain other options!

Discovered while working on the bulk renderer.