fix(windows): Splash button sizes
Fixes #4265.

This adjusts the button sizes of the splash screen to accommodate longer
strings such as found in the Khmer translation. We could go to a more
flexible layout model to allow for the buttons to stretch as needed, but
that's a lot more work and I don't really want to try for that right
now.