fix(windows): Ignore Access Denied error creating task
Fixes #4116.
Fixes KEYMAN-WINDOWS-5T.

If we get an 'access denied' error, we won't report it to Sentry. This
error is of minor consequence as it arises in normal configurations only
if the task was already present, created by an elevated instance of
kmshell under the same user name. In that situation, the task works
fine; it just cannot be updated by a non-elevated instance of kmshell.
If there are other reasons for receiving access denied, we won't be able
to distinguish them anyway, so the best response is to suppress this
message.

An alternative would have been to either (a) avoid creating the task as
an elevated user, or (b) only ever create the task as an elevated user.
However, this adds extra unnecessary complexity and means that the task
may not always be present when we want it to be.